# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mass Finder API",
        "description": "API para busca de paróquias e horários de missa",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{baseUrl}/api",
            "description": "API Base URL",
            "variables": {
                "baseUrl": {
                    "default": "http://localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/admin/parishes": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Listar todas as paróquias",
                "operationId": "cdedd0bf5de4ed2fcf9a509dc5c30ef6",
                "responses": {
                    "200": {
                        "description": "Lista de paróquias",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AdminParish"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/parishes/{parish}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Exibir paróquia por ID",
                "operationId": "2fd90f6f1f4b7630233a7a02c43b5b55",
                "parameters": [
                    {
                        "name": "parish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paróquia com horários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AdminParishDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Atualizar paróquia",
                "operationId": "69664fa5333e861fb4b864d3cbe516ba",
                "parameters": [
                    {
                        "name": "parish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AdminUpdateParishRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paróquia atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AdminParishDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Remover paróquia",
                "operationId": "3d17ecdb0bb7b4ab8c66d8d5376043ea",
                "parameters": [
                    {
                        "name": "parish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paróquia removida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/parishes/{parish}/approve": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Aprovar paróquia",
                "operationId": "939b5493e5ddc78d5cbbed8a6bf916bf",
                "parameters": [
                    {
                        "name": "parish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paróquia aprovada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AdminParishDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/parishes/{parish}/reject": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Rejeitar paróquia",
                "operationId": "c644381e6ad0e8a9a9b8a3e2f4f2fcaf",
                "parameters": [
                    {
                        "name": "parish",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RejectParishRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paróquia rejeitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AdminParishDetail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Paróquia não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericMessageResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidationResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MassSchedule": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "parish_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "day_of_week": {
                        "description": "0=Domingo, 1=Segunda, ..., 6=Sábado",
                        "type": "integer",
                        "maximum": 6,
                        "minimum": 0
                    },
                    "time": {
                        "type": "string",
                        "example": "08:00"
                    },
                    "notes": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AdminParish": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Paróquia São José"
                    },
                    "cnpj": {
                        "type": "string",
                        "example": "12.345.678/0001-90"
                    },
                    "neighborhood": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ]
                    },
                    "rejection_reason": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AdminParishDetail": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Paróquia São José"
                    },
                    "cnpj": {
                        "type": "string",
                        "example": "12.345.678/0001-90"
                    },
                    "neighborhood": {
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ]
                    },
                    "rejection_reason": {
                        "type": "string",
                        "nullable": true
                    },
                    "mass_schedules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MassSchedule"
                        }
                    }
                },
                "type": "object"
            },
...
            "AdminUpdateParishRequest": {
                "required": [
                    "name",
                    "cnpj"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Paróquia São José"
                    },
                    "cnpj": {
                        "type": "string",
                        "maxLength": 18,
                        "example": "12.345.678/0001-90"
                    },
                    "neighborhood": {
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "maxLength": 500,
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "nullable": true
                    },
                    "rejection_reason": {
                        "type": "string",
                        "maxLength": 1000,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "RejectParishRequest": {
                "required": [
                    "rejection_reason"
                ],
                "properties": {
                    "rejection_reason": {
                        "type": "string",
                        "maxLength": 1000,
                        "example": "Documentação incompleta"
                    }
                },
                "type": "object"
            },
            "GenericMessageResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Operação realizada com sucesso."
                    }
                },
                "type": "object"
            },
            "ErrorValidationResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token obtido via POST /api/auth/login",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}